requirements.txt
alembic==0.9.3
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
click==6.7
DateTime==4.2
Django==1.11.3
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
get==0.0.0
gunicorn==19.7.1
idna==2.5
image==1.5.7
itsdangerous==0.24
Jinja2==2.9.6
jsonify==0.5
Mako==1.0.7
MarkupSafe==1.0
numpy==1.13.0
olefile==0.44
pandas==0.20.2
Pillow==4.2.1
post==0.0.0
psutil==5.2.2
psycopg2==2.6.1
public==0.0.0
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.2
query-string==0.0.0
request==0.0.0
requests==2.18.1
setupfiles==0.0.50
six==1.10.0
SQLAlchemy==1.1.11
urllib3==1.21.1
virtualenv==15.1.0
Werkzeug==0.12.2
zope.interface==4.4.2
